+++ stwo_cairo_prover/crates/prover/src/cairo_air/debug_tools/assert_constraints.rs
@@ -13,7 +13,7 @@ use crate::cairo_air::air::{CairoComponents, CairoInteractionElements};
 use crate::cairo_air::debug_tools::mock_tree_builder::MockCommitmentScheme;
 use crate::cairo_air::opcodes_air::OpcodeComponents;
 use crate::cairo_air::preprocessed::PreProcessedTrace;
-use crate::cairo_air::witness::CairoClaimGenerator;
+use crate::witness::cairo::CairoClaimGenerator;
 
 pub fn assert_component<E: FrameworkEval + Sync>(
     component: &FrameworkComponent<E>,
