+++ stwo_cairo_prover/crates/prover/src/cairo_air/poseidon/mod.rs
@@ -1,3 +1,2 @@
 pub mod air;
 pub mod const_columns;
-pub mod deduce_output;
