+++ stwo_cairo_prover/crates/prover/src/components/prelude.rs
@@ -21,15 +21,15 @@ pub mod proving {
     pub use stwo_prover::core::poly::circle::{CanonicCoset, CircleEvaluation};
     pub use stwo_prover::core::poly::BitReversedOrder;
 
-    pub use crate::cairo_air::poseidon::deduce_output::{
-        PackedCube252, PackedPoseidon3PartialRoundsChain, PackedPoseidonFullRoundChain,
-        PackedPoseidonRoundKeys,
-    };
     pub use crate::cairo_air::preprocessed::Seq;
     pub(crate) use crate::cairo_air::relations;
     pub use crate::components::utils::*;
     pub use crate::witness::fast_deduction::blake::{
         BlakeG, BlakeRound, BlakeRoundSigma, TripleXor32,
+    };
+    pub use crate::witness::fast_deduction::poseidon::{
+        PackedCube252, PackedPoseidon3PartialRoundsChain, PackedPoseidonFullRoundChain,
+        PackedPoseidonRoundKeys,
     };
 }
