+++ stwo_cairo_prover/crates/prover/src/witness/cairo.rs
@@ -11,14 +11,14 @@ use super::range_checks::{RangeChecksClaimGenerator, RangeChecksInteractionClaim
 use crate::cairo_air::air::{
     CairoClaim, CairoInteractionClaim, CairoInteractionElements, PublicData,
 };
-use crate::cairo_air::poseidon::air::{
-    PoseidonContextClaimGenerator, PoseidonContextInteractionClaimGenerator,
-};
 use crate::components::memory::{memory_address_to_id, memory_id_to_big};
 use crate::components::utils::TreeBuilder;
 use crate::components::{
     verify_bitwise_xor_4, verify_bitwise_xor_7, verify_bitwise_xor_8, verify_bitwise_xor_9,
     verify_instruction,
+};
+use crate::witness::components::poseidon::{
+    PoseidonContextClaimGenerator, PoseidonContextInteractionClaimGenerator,
 };
 
 /// Responsible for generating the CairoClaim and writing the trace.
