+++ stwo_cairo_prover/crates/prover/src/witness/components/mod.rs
@@ -1,2 +1,1 @@
-pub mod blake;
 pub mod poseidon;
