+++ stwo_cairo_prover/crates/prover/src/witness/mod.rs
@@ -1,6 +1,7 @@
 pub mod blake_context;
 pub mod builtins;
 pub mod cairo;
+mod components;
 pub mod fast_deduction;
 pub mod opcodes;
 pub mod range_checks;
